Add overloads for IPubSub.PublishAsync and .SubscribeAsync to pass data into the configuration callbacks
When we do

	pubSub.PublishAsync(message, config => ...)

if we need to reference any data from the surrounding context it will allocate a closure on the heap. Multiply this times however many messages you publish.
Compare to:

    pubSub.PublishAsync(message, contextualData, static (config, data) => ...)

Here we can use a static callback method and avoid allocating a closure.
I have done the same thing for .SubscribeAsync, though I suspect Subscribes happen far less often so the big win is the change to .PublishAsync